more details for AdministratorClient

+++ src/main/java/alessio_la_greca_990973/server/forclient/TaxiStatisticWithTimestamp.java
@@ -4,20 +4,20 @@ import java.util.List;
 
 public class TaxiStatisticWithTimestamp extends TaxiStatistic{
 
-    private long timestamp;
+    private double timestamp;
     private List<Double> pollutionAverages;
 
-    public TaxiStatisticWithTimestamp(int kilometers, int rides, List<Double> pollutionAverages, int batteryLevel, long timestamp){
+    public TaxiStatisticWithTimestamp(int kilometers, int rides, List<Double> pollutionAverages, int batteryLevel, double timestamp){
         super(kilometers, rides, 0D, batteryLevel);
         this.timestamp = timestamp;
         this.pollutionAverages = pollutionAverages;
     }
 
-    public long getTimestamp() {
+    public double getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(long timestamp) {
+    public void setTimestamp(double timestamp) {
         this.timestamp = timestamp;
     }
